git commit -m "fix : nullable, status 변경"

+++ Omniscient-Back/src/main/java/com/omniscient/omniscientback/manager/faq/repository/FaqRepository.java
@@ -4,5 +4,5 @@ package com.omniscient.omniscientback.manager.faq.repository;
 import com.omniscient.omniscientback.manager.faq.model.Faq;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FaqRepository extends JpaRepository<Faq, Long> {
+public interface FaqRepository extends JpaRepository<Faq, Integer> {
 }
